requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client==1.7.3
sentence-transformers==2.3.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-frontmatter==1.1.0
langchain==0.1.6
langchain-community==0.0.19
openai==1.10.0
tiktoken==0.5.2
httpx<0.27.0
